refactor: shareOfCheckout + test
