+++ Test/Unit/Cron/ShareOfCheckoutTest.php
@@ -3,9 +3,9 @@
 namespace Alma\MonthlyPayments\Test\Unit\Cron;
 
 use Alma\MonthlyPayments\Cron\ShareOfCheckout;
-use Alma\MonthlyPayments\Helpers\DateHelper;
+use Alma\MonthlyPayments\Helpers\ShareOfCheckout\DateHelper;
 use Alma\MonthlyPayments\Helpers\Logger;
-use Alma\MonthlyPayments\Helpers\ShareOfCheckoutHelper;
+use Alma\MonthlyPayments\Helpers\ShareOfCheckout\ShareOfCheckoutHelper;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -15,6 +15,18 @@ class ShareOfCheckoutTest extends TestCase
      * @var MockObject|Logger
      */
     private $logger;
+    /**
+     * @var ShareOfCheckoutHelper|MockObject
+     */
+    private $shareOfCheckoutHelper;
+    /**
+     * @var DateHelper|MockObject
+     */
+    private $dateHelper;
+    /**
+     * @var ShareOfCheckout
+     */
+    private $shareOfCheckoutCron;
 
     protected function setUp(): void
     {
@@ -37,5 +49,24 @@ class ShareOfCheckoutTest extends TestCase
         $this->shareOfCheckoutHelper->expects($this->once())
             ->method('shareOfCheckoutIsEnabled')
             ->willReturn(false);
+        $this->shareOfCheckoutHelper->expects($this->never())
+            ->method('getShareOfCheckoutEnabledDate');
+        $this->shareOfCheckoutCron->shareDays();
+    }
+    public function testShareDaysWithShareEnable()
+    {
+        $shareOfCheckoutEnabledDate = '2022-04-24';
+        $lastUpdateDate = '2022-04-20';
+
+        $this->shareOfCheckoutHelper->expects($this->once())
+            ->method('shareOfCheckoutIsEnabled')
+            ->willReturn(true);
+        $this->shareOfCheckoutHelper->expects($this->once())
+            ->method('getShareOfCheckoutEnabledDate')
+            ->willReturn($shareOfCheckoutEnabledDate);
+        $this->shareOfCheckoutHelper->expects($this->once())
+            ->method('getLastUpdateDate')
+            ->willReturn($lastUpdateDate);
+        $this->shareOfCheckoutCron->shareDays();
     }
 }
